requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-chess==1.999
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.65.0
pydantic>=2.0.0
